Randomize data set
This allows steps/epochs to be set higher without
causing learning to go awry.

Since b is low, and learning rate is probably even lower,
when the model is exposed to non-randomized data with initial
b of zero, it quickly attunes to the value of y that are close
to b, but when y grows much bigger it cannot cope/learn fast enough.